Fix FS-3904 use a message to shutdown the module
For some reason the running flag being set on the thread initiating the
shutdown did not propogate to the run loop thread. I couldn't find any
memory barrier functions in APR or the FS API so I am using 0MQ to send
a termination request message to the run loop thread.