Harmonize debian nightly version with print_git_revision
Debian nightlies created with debian/util.sh now have the version set
in the same format as Makefile.am produces.  Unlike Makefile.am,
however, we show the orig packaging date rather than the date of the
latest commit, as the latter could legitimately regress and we need
this to always increment.