Ensure that flite gets configured before building
Without this change, we would only configure flite if its Makefile did
not exist.  However in the tarball for flite (v1.5.4), there is a
Makefile even before you run configure.  So the only reason this
worked at all was because make was calculating the rules to run before
downloading and extracting flite.

If you build flite without configuring it first, you get a flite with
ALSA symbols that doesn't load in FS.

In discussion with the flite guys, they indicated that the presence of
the Makefile in the tarball is indeed intended.

This change has the side-effect of fixing mod_flite in the Debian
packages.

FS-5873 --resolve