Prevent -lfoo lib flags from becoming dependencies
Previously we were including the output of `apr-1-config/apu-1-config
--libs` into CORE_LIBS.  This seemed sensible as CORE_LIBS is added in
the right places to various _LDADD and _LIBADD variables.  However we
also use the value of CORE_LIBS as dependencies in the makefile.
Astonishingly this doesn't seem to cause any issues on most
platforms (or we would have noticed sooner), but it does break things
on Mac.

Thanks-to: Seven Du <dujinfang@gmail.com>

FS-5956 --resolve