Add vm-domain-storage-dir param
This replaces the semantics of the storage-dir parameter without the
deprecation notice.  The behavior of storage-dir never should have
been deprecated, though it was probably correct to deprecate the name
itself.

Rather than having both vm-storage-dir and vm-domain-storage-dir it
might be nice if the config system could just notice at what level it
was defined and do the right thing.  Indeed, this is why I'd left
things as-is for so long.

But as Master Foo famously said, "which one will reach the other side
of the river: the one who dreams of a raft, or the one that hitchhikes
to the next bridge?"

FS-5329