Send silent packets when idle with SRTP
Originally we did the same thing with SRTP that we do without SRTP,
which is to simply not send packets when e.g. sleep is called.

At commits d63323977fa611b141441f12af9a94ec19b5f829 and
5259814aee16ede974456490a79e8a98de1d6d2e we enabled sending silence
packets with comfort noise when SRTP is active.  We appear to have
done this for interop purposes; many devices can't handle gaps in the
stream of SRTP packets.

But our current comfort noise implementation doesn't take the codec
rate into account (FS-6291), so on 16kHz codecs the constant we chose
created an annoying level of static between sound file playback.

With this commit we preserve the sending of SRTP packets during idle
periods, but make those packets completely silent.

Thanks-to: Anthony Minessale <anthm@freeswitch.org>

FS-5053 --resolve