Pull mod_v8 out of the default build
We love mod_v8 and Peter for making it work so well and in so many
places.

But it still doesn't build on as many platforms as the rest of
FreeSWITCH, and it requires using getlib on a huge library while we're
currently trying to pull libraries out of tree and reduce our
dependence on getlib.  We need to work on getting distributions to
push updated versions of libv8.