Prevent opus lib from re-configuring on every make
We were using the libopus directory itself rather than a file within
it as a makefile target.  This is a bad idea as make will see the
target as updated whenever files are created in or removed from the
directory.  Since running configure creates and removes files, the
target will always appear dirty.

The effect was that when mod_opus was enabled in FS make would always
cause libopus to be reconfigured.

We'll target instead the configure file in libopus as that's the first
thing we need anyway.