Fix source fraction always 0 in RTCP events
Without the value for source fraction, applications relying on RTCP events for making changes to FS behaviour or even for logging get false information.

With this change the value for source fraction is passed along in RTCP events correctly.

To my current understanding, as the value for fraction in the RTCP packet is represented by 8 bits according to the spec, calling `ntohl` on it will always zero it out. Fixed by removing the call.

FS-7203 #resolve