FS-6758 and other related behavior, also change to answer behavior to mimic what users tend to expect from cisco phones.
Updatd SKINNY on-hook action to hang up all calls on a device, except those in a short list of call states (or perform a blind transfer).
Added a hook after completing the hangup operation to start ringing if there is an inbound call active on the device.

Reorder received calls on skinny