FS-8978: [mod_redis] Fix limit counter not decrementing on hangup
The limit counter value is not automatically decremented in Redis when
the call is hung up. This results in it only ever counting up and the
limit being eventually hit.

The linked list is moved on too early. When `limit_pvt` is checked, it
is `NULL`. This means the `decr` command is never sent to Redis.